Added new append method to LinkedList class
This new code allows me to append a new node
to the end of our LinkedList. The logic behind this,
given the LinkedList was not null, was to point the
next variable to tail (the last item in the current LL)
to the new node and then point tail variable to the new node.