Add method to remove the last node in a LinkedList
This method was added to give us the ability to remove the last
item in the LinkedList. Logic behind this is we need two Node
variables (temp & pre) that point to head. As we iterate through
the LinkedList, we set pre equal to temp and set temp equal to
temp.next. Once temp.next is equal to null, we break out of the loop,
set tail equal to pre, and set tail.next equal to null to break off the
last item. Finally, we return that temp node as the node that needs to
be removed.