Add prepend method for LinkedList
This method gives us the ability to prepend a new node to the
front of our LinkedList. Logic behind this is we need to set
newNode.next = head, then set head = newNode. This will add
our new node to the beginning of our LinkedList. Finally we
increment the length by 1