Add method to remove first node in LinkedList
The removeFirst method gives us the ability to remove
the first node in the LinkedList. Logic for this is to
return null if the length of the LinkedList is 0. Otherwise
we create a new 'temp' node that points to head to track the
first node. We then point head to the next node in the LinkedList
and set temp.next to null to break that first node away from the list.
Finally, we decrease the length by 1. If in the case there was only
one item in the LinkedList, we then check again if length is 0. If
it is, we then set tail equal to null.