Add method to get node at index
In this method, we will return the node at a specified value.
First, we check our IndexOutOfBounds parameters. Then we create
a new Node to keep track of where we are in the LinkedList.
Looping through the LinkedList, once our temp value is at our specified
index, we return that Node