Add method to set the value of an existing node
This method gives us the ability to set a new value
on an existing node. We pass in the index we want to change
and the value to update that node to. We can utilize the get
method by passing in the index we want to update and if that
is not null, we set the temp.value equal to the value we pass in