Add insert method to add new node
This method gives us the ability to insert a new node in our
LinkedList. If index is 0, we just call our existing prepend method;
if index is equal to the length of the LinkedList, we call our
existing append method. To insert in the middle, we need to have a
temp node that points to the node just before our desired index. We
then point our new node to temp.next, then point temp.next to the
new node.