Add method to remove a node at specified index
This method gives us the ability to remove a specific node
at the given index. Logic is similar to that of insert. We
check for our out of bounds issue and return null. If index
is equal to zero, we return our removeFirst method. If index
is equal to length - 1, we return our removeLast method. Otherwise,
we create two node variables, temp and prev, pointing to index
to be removed and index just before, respectfully. We point prev.next
to temp.next and then point temp.next to null