Add reverse method to flip the order of our list
This method is a in-place solution to flip the order
of the LinkedList. No new list or copy of the list is created.
Logic for this is to create a temp variable, set head equal
to tail and tail equal to temp. This flips the order of the first
and last nodes in the list. Finally, we create an after and before
variable. As we loop through the list, we essentially just flip
the order of the pointers for the middle nodes