Coding excercise to get kth node from end
In this excercise, this new method was created to find the kth
node from the end of the LinkedList. To do this, we use two pointers
with one moving twice as fast as the other. Once the 'fast' pointer
reaches the end, the 'slow' pointer will be pointing to the kth value