Add function to partition the LinkedList
This method takes in an integer that we use as we loop through
the list. If the value of the current node is less than x,
we move that value to the first partition. If it is greater than
x, we move it to the second partition. We then concat those partitions
when the loop finishes